# docs/api-specification.yaml
openapi: 3.1.0
info:
  title: AccountabilityAtlas Video Service API
  version: 1.0.0
  description: |
    Core content management service for AccountabilityAtlas.
    Manages video records, YouTube metadata integration, amendment categorization,
    and video-location associations.

    ## Access Patterns
    - **Public (via Gateway)**: `https://api.accountabilityatlas.com/api/v1/videos/...`
    - **Internal (service-to-service)**: `http://video-service:8082/videos/...`

    The API Gateway adds the `/api/v1` prefix for public requests.
    Services calling directly use paths without the version prefix.
  contact:
    name: AccountabilityAtlas Team
  license:
    name: MIT
    identifier: MIT
servers:
  - url: http://localhost:8082
    description: Local development (direct)
  - url: http://video-service:8082
    description: Internal service mesh (Kubernetes/ECS service discovery)
  - url: https://api.accountabilityatlas.com/api/v1
    description: Production (via API Gateway)
...
security:
  - bearerAuth: []
paths:
  /videos:
    get:
      operationId: listVideos
      summary: List videos with filtering and pagination
      description: |
        Returns a paginated list of videos. By default, only APPROVED videos are returned.
        Supports filtering by amendments, participants, submitter, and status.
      tags: [Videos]
      security: []
      parameters:
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/VideoStatus'
          description: "Filter by status (default: APPROVED)"
        - name: amendments
          in: query
          schema:
            type: array
            items:
              $ref: '#/components/schemas/Amendment'
          style: form
          explode: true
          description: Filter by amendments (OR logic)
        - name: participants
          in: query
          schema:
            type: array
            items:
              $ref: '#/components/schemas/Participant'
          style: form
          explode: true
          description: Filter by participants (OR logic)
        - name: submittedBy
          in: query
          schema:
            type: string
            format: uuid
          description: Filter by submitter user ID
        - name: page
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
          description: Page number (0-indexed)
        - name: size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Page size
        - name: sort
          in: query
          schema:
            type: string
            enum: [createdAt, videoDate]
            default: createdAt
          description: Sort field
        - name: direction
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
          description: Sort direction
      responses:
        '200':
          description: Paginated list of videos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VideoListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'

    post:
      operationId: createVideo
      summary: Submit a new video
      description: |
        Submit a new video for moderation. The service fetches metadata from YouTube.
        Videos start in PENDING status unless submitter is TRUSTED or higher.
      tags: [Videos]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateVideoRequest'
      responses:
        '201':
          description: Video submitted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VideoDetail'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Video already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: "VIDEO_EXISTS"
                message: "This YouTube video has already been submitted"
                traceId: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
        '422':
          description: YouTube video not found or unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: "YOUTUBE_VIDEO_UNAVAILABLE"
                message: "The YouTube video is private, deleted, or does not exist"
                traceId: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
        '429':
          $ref: '#/components/responses/RateLimited'
  /videos/preview:
    get:
      operationId: previewVideo
      summary: Preview YouTube video metadata
      description: |
        Returns YouTube metadata for the given URL without creating a video record.
        Useful for previewing video details before submission.
        Also indicates whether the video has already been submitted.
      tags: [Videos]
      security: []
      parameters:
        - name: youtubeUrl
          in: query
          required: true
          schema:
            type: string
          description: Full YouTube URL to preview
      responses:
        '200':
          description: YouTube video metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VideoPreview'
        '400':
          description: Invalid YouTube URL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: "INVALID_YOUTUBE_URL"
                message: "Invalid YouTube URL: https://example.com/not-a-video"
                traceId: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
        '422':
          description: YouTube video not found or unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: "YOUTUBE_VIDEO_UNAVAILABLE"
                message: "The YouTube video is private, deleted, or does not exist"
                traceId: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
  /videos/extract:
    get:
      operationId: extractVideoMetadata
      summary: Extract metadata from YouTube video using AI
      description: |
        Uses AI to analyze a YouTube video's title and description to suggest
        amendments, participants, incident date, and location.
        Requires authentication to prevent abuse.
      tags: [Videos]
      parameters:
        - name: youtubeUrl
          in: query
          required: true
          schema:
            type: string
          description: Full YouTube URL to extract metadata from
      responses:
        '200':
          description: AI-extracted metadata suggestions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VideoMetadataExtraction'
        '400':
          description: Invalid YouTube URL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: "INVALID_YOUTUBE_URL"
                message: "Invalid YouTube URL: https://example.com/not-a-video"
                traceId: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          description: YouTube video not found or unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: "YOUTUBE_VIDEO_UNAVAILABLE"
                message: "The YouTube video is private, deleted, or does not exist"
                traceId: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
        '503':
          description: AI service unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: "AI_SERVICE_UNAVAILABLE"
                message: "AI metadata extraction is temporarily unavailable"
                traceId: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
  /videos/{id}:
    get:
      operationId: getVideo
      summary: Get video details
      description: Returns full video details including locations and submitter info.
      tags: [Videos]
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Video details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VideoDetail'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      operationId: updateVideo
      summary: Update video metadata
      description: |
        Update video amendments, participants, or date.
        YouTube metadata (title, description, etc.) cannot be changed.

        **Authorization:**
        - Owner: Can update only if video status != APPROVED
        - Moderators/Admins: Use moderation-service endpoints instead
      tags: [Videos]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateVideoRequest'
      responses:
        '200':
          description: Video updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VideoDetail'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      operationId: deleteVideo
      summary: Delete a video
      description: |
        Soft-delete a video. Sets status to DELETED; video is hidden but data retained.

        **Authorization:**
        - Owner: Can delete only if video status != APPROVED
        - Moderators/Admins: Use moderation-service endpoints instead
      tags: [Videos]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Video deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /videos/user/{userId}:
    get:
      operationId: getVideosByUser
      summary: Get videos submitted by a user
      description: |
        Returns paginated list of videos submitted by the specified user.
        Only APPROVED videos are shown unless viewing own submissions.
      tags: [Videos]
      security: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/VideoStatus'
          description: Filter by status (only own submissions can see non-APPROVED)
        - name: page
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: User's videos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VideoListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /videos/{id}/locations:
    get:
      operationId: getVideoLocations
      summary: Get locations associated with a video
      description: Returns all locations linked to the video with full location details.
      tags: [VideoLocations]
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Video ID
      responses:
        '200':
          description: Video locations
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VideoLocationsResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    post:
      operationId: addVideoLocation
      summary: Add a location to a video
      description: |
        Associate an existing location with the video.
        If isPrimary is true, any existing primary location is demoted.

        **Authorization:**
        - Owner: Can add locations only if video status != APPROVED
        - Moderators/Admins: Use moderation-service endpoints instead
      tags: [VideoLocations]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Video ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddVideoLocationRequest'
      responses:
        '201':
          description: Location added to video
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VideoLocation'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Video or location not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: "NOT_FOUND"
                message: "Video or location not found"
                traceId: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
        '409':
          description: Location already associated with video
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: "LOCATION_ALREADY_LINKED"
                message: "This location is already associated with the video"
                traceId: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT access token from user-service.
        Obtain via `/auth/login` or `/auth/oauth/{provider}`.
  schemas:
...
    Amendment:
      type: string
      enum: [FIRST, SECOND, FOURTH, FIFTH, FOURTEENTH]
      description: |
        Constitutional amendment categories:
        - `FIRST`: Freedom of speech, press, assembly
        - `SECOND`: Right to bear arms
        - `FOURTH`: Protection against unreasonable search/seizure
        - `FIFTH`: Right against self-incrimination
        - `FOURTEENTH`: Due process and equal protection
    Participant:
      type: string
      enum: [POLICE, GOVERNMENT, BUSINESS, CITIZEN, SECURITY]
      description: Types of participants in the recorded encounter
    VideoStatus:
      type: string
      enum: [PENDING, APPROVED, REJECTED, DELETED]
      description: |
        Video moderation status:
        - `PENDING`: Awaiting moderation review
        - `APPROVED`: Published and visible
        - `REJECTED`: Failed moderation
        - `DELETED`: Removed by owner or moderator

    # Core domain objects
...
    VideoListResponse:
      type: object
      required: [content, page, totalElements, totalPages]
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/VideoSummary'
        page:
          type: integer
          minimum: 0
          description: Current page number (0-indexed)
        size:
          type: integer
          minimum: 1
          maximum: 100
          description: Page size
        totalElements:
          type: integer
          minimum: 0
          description: Total number of videos matching query
        totalPages:
          type: integer
          minimum: 0
          description: Total number of pages
...